requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
codetiming==1.4.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
